Quoting :DOC note-8-6-book:
  [Read-string] and read-string-light-fn now take a required package
  argument, which can be nil to represent the [current-package], thus
  providing the previous behavior.

Thanks to Eric Smith for requesting this enhancement.  I agree with
him that it's potentially risky for the behavior to depend on the
current package.  That said, I just added a nil argument in calls of
read-string in the books, to preserve the current behavior.  Those who
use read-string may want to consider passing in a non-nil package
name.