[ATC] Add rule to some generated hints.
This is for the theorems about the results of ACL2 functions. Besides saying
that the function return values of the expected C types, these theorems also say
that the values are not `nil`, which seems a bit of an odd thing to be needed,
but it did come up in symbolic execution proofs. These are normally proved via
suitable type prescription rules. We added more of these type prescription rules
to the generated hints for these theorems.

Thanks to Eric Smith for running an example that failed in proof generation,
revealing the issue.