[ATC] Improve handling of external objects.
When external objects (i.e. `defobject`s) have initializers, do not put the
initializer in the `.h` file, which may cause redefinition errors in the C
compiler when the generated files are compiled with others. Instead, just put
the declaration without initializer in the `.h` file, and then the one with the
initializer in the `.c` file.