[ATC] Use `void` for no function parameters.
The reason is explained in the added dev doc. For now we do this just by
modifying the pretty-printer, but in the future we should capture the `void`
parameter list explicitly in the abstract syntax.

Thanks to Grant Jurgensen for raising the issue of `void` parameter list for
ATC-generated code.