[ATC] Improve propagation of the `proofs` flag.
This does not make a difference currently, but it will make a difference as we
generate additional modular proofs. The plan is to add support for modular proof
generation for C constructs incrementally: the flag starts `t` but it becomes
`nil` when encountering a construct for which modular proof generation is not
supported yet.

This commit properly propagates the flag for this purpose.