[Std/util] Add new `defirrelevant` macro.
This automates some of the boilerplate involves in defining a value of a given
type where the exact definition of the value is irrelevant, as it is used as a
dummy value of the right type. A use case is when a function returns multiple
results, some of which are irrelevant when others have certain values (e.g. the
`val` of an error triple may be irrelevant when `erp` is non-`nil`), but we
still want to return a result of the right type, avoiding conditional return
theorems. This is mainly meant for types where constructing any legal value
takes more than a short phrase like `nil` for symbols and lists, or like `""`
for strings, or like `0` for numbers; for example, one can use this to define an
irrelevant value of a sum or product fixtype.

There is an unrelated `defirrelevant` in the COI library, but it is in a
different package. This new one is in the `ACL2` package, but it could be moved
to `STD` if desired.