[Std/util] Improve error-value tuples.
Use `macrolet` instead of `flet` (via `b*`'s `fun`) for `reterr`.

This provides support for using non-ground terms for the irrelevant values to
return in case of error, which in particular allows results that are stobjs,
such as state.

Thanks to Matt Kaufmann for adding `macrolet` to ACL2, which has made this
enhancement to the error-value tuple utilities possible.