[C] Tighten the dynamic semantics.
Ensure that, when we evaluate struct member expressions (by value or by pointer)
as pure expressions, the member value is not an array. This is because, to
properly handle that, we need a more elaborate model of expression evaluation,
which we plan to have later.

This was not an issue for proofs generated by ATC, where struct array members
are never used in pure expressions.