[ATC] Improve theorem generation.
Also return the name of the correctness theorem for a pure expression, in the
data structure returned by the generation functions. (A similar extension will
be done to the analogous data structures for the generation of other expressions
and statements.) Use the new slot to return the name of the correctness theorem
for a constant expression.