[ATC] Generate some double-checking assertions.
These are to address (future occurrences of something like) the issue that was
fixed by the previous commit. Now we generate assertions saying that all the
expected top-level theorems have been generated. The assertions only check that
the names indeed refer to theorems; it does not look at the theorems. This
suffices to catch issues like the one fixed in the previous commit.