[helpers] Improve advice tool.
Handle the case where a recommended library brings in an event with the same name as the theorem we're trying to prove.  For now, in that case, we always use a temporary theorem (with a non-clashing name) which is proved by locally including the library and then is used to prove the desired theorem.