[ABNF] Extend `deftreeops`.
For each rule name, generate a theorem to do case splits based on the
alternative. Currently this takes long if there are too many alternatives, so
for now we just cap the generation of this theorem to 10 alternatives or less.
We will look into more nuanced hints, or proof builder instructions, to avoid
the slow speed. We already avoid preprocessing, which was causing slow speed
already.