[ATC] Remove support for `not`.
Before this commit, ATC supported two ways to represent the C logical negation
operator `!`: as `(not ...)`, and as `(boolean-from-sint (lognot-... ...))`.
This was a mistake, in retrospect: there should be just one way to represent
each C construct, as ATC's approach is to shift the burden of putting things
into the appropriate C representation to APT transformations outside ATC.

While the C operators `&&` and `||` need to be represented via ACL2's `and` and
`or` because of non-strictness, there is no such need for `!` and `not`.