[ATC] Extend modular proof generation.
Add code to generate proofs for conditional expressions. The generated event is
not actually submitted because there is another problem to solve first, namely
the fact that guard theorems are still in terms of `if`, while the modular
theorems are in terms of `if*`: this discrepancy is causing failures in the
examples. We will address this issue next.