[built-ins] Add a library for built-ins.
This library collects and categorizes built-in functions, axioms, theorems, and
other events. It also XDOCuments the built-in axioms and theorems, categorized
by rule classes and by topic. Finally, it provides macros to disable built-in
functions and rules -- currently just one macro to disable all the built-in
logic-mode functions.

The purpose is to see and manipulate more easily (than browsing the source code)
the built-ins that affect the theory in which proofs are carried out.