[ATC] Speed up certification of a file.
After disabling all the built-in logic-mode functions in
`expression-generation.lisp`, and only enabling the ones needed for the proofs
to go through where needed, the certification of this file dropped from 45 to 35
seconds on a machine.

The enabling of `alistp` in the fixtype definitions can be probably "moved" into
the fixtype definition macros themselves.