[ATC] Slight speed up in certification.
Try disabling two built-in theorems. The file certifies without modification, so
they are not needed. These have been found as useless in accumulated persistence
several times in the past. For this particular file, there seems to be no
significant speed up though. But we leave the disabling there for now.