[ATC] Speed up the certification of a file.
This is also done by disabling all the built-in logic-mode functions, and making
some other adjustments. For this case the speed-up is less than the other two
files: about 15% less on a Mac, and about 5% less on a Linux PC.

Also re-generate the useless runes file, to prevent certification from failing.