[ATC] Extend modular proof generation.
Generate proofs for non-strict `&&` expressions.

However, currently we stop modular proof generation just after generating the
proofs for `&&`, because otherwise some of our current uses of ATC fail, due to
ACL2's replacement of `<term>` with `nil` when there is a hyp `(not <term>)`,
which prevents theorems for certain subexpressions to apply.

This may necessitate introducing a synonym `not*` of `not` that is not treated
specially by ACL2 (similarly to the use of `if*` as a more controllable synonym
of `if`).