[PFCS] Improve the notion of system.
Instead of defining a (PFCS) system as a list of definitions, we now define it
as consisting of a list of definitions and a list of constraints. This is a more
natural notion, since a system is something that should express constraints that
may have solutions; the definitions provide a context for the constraints. Note
that the constraints could be just one that is a call of a named relation on its
parameters.