[ATC] Add some execution rules.
These are used for the new modular proof generation. We want to avoid
introducing `if` from right sides of rules. So we split some of the execution
rules, used in the large symbolic execution, into two rules, one for the 'true'
case and one for the 'false' case: the two rules will be used in different hints
for different proofs (the ones relative to each branch).