[Seq] Improve `:s=` and `:w=`.
Instead of returning the original stream if the `mbt` check fails, return the
remaining stream from the call that fails the `mbt` check. The advantage is that
often the stream processing functions fix the stream to its type, and therefore
unconditionally return a remaining stream of the right type. However, if `:s=`
and `:w=` return the initial stream, which is not fixed to the right type, then
function using these operators need conditional return theorems. Note that the
exact stream returned if the `mbt` check fails does not matter in guard-verified
code, because it is provably never returned. I'm not sure if there are good use
cases for returning the original stream instead, when the code is not
guard-verified (e.g. in program mode).