[ABNF] Improve grammar parser implementation.
Use Seq's `:s=` to ensure the termination of the mutually recursive parsing
functions, instead of manually doing the error checking and propagation and the
`mbt` tests. The parser was initially not written with `:s=` because the Seq
documentation said that the `mbt` check is on the `acl2-count`, while the
parsing functions needed one on `len`. However, it turns out that the `:s=`
`mbt` check is on the length; the Seq documentation has been corrected in a
previous commit.