[ATC] Strengthen some generated theorems.
For the generated modular proofs for non-call expressions, which involve limit
hyps, strengthen the part about the type of the expression by using an
implication without limit hyps for that part. Also avoid the computation state
bindings for that part, even though it does not appear in the generated rule.