[built-ins] Properly exclude functions in the file.
Now that the file starts with two functions, we need to exclude two of them
(with `cddr`) instead of just one (with `cdr`) from the list of function names
in reverse chronological order.

Also add comments explaining why the `set-compile-fns` form does not need to be
taken into account for this.

Also align some text more consistently.