[ATC] Use new `test*` instead of `hide`.
After another issue came up with `hide` (in the course of making an orthogonal
improvement to modular proof generation), it seems best to avoid using `hide`
and instead use a new custom wrapper `test*`, which was in fact the original
plan. This gives us more assurances of being able to control things, and also
simplifies some generated hints, because to expand `test*` (in the relatively
few places where it must be opened) we just need to enable it, as opposed to
using an `:expand` hint for `hide`.