[ATC] Removed unneeded code.
It turns out that the "star" version of the guard theorem (i.e. the guard
theorem with every `if` replaced with `if*`) is not needed. In fact, we need to
use (as we do) the actual guard theorem with `if` for things to work
properly. This commit removes the code related to the star version of the guard
theorem.