[ATC] Slightly simplify screen printing.
Remove support for the `:print nil` option. The idea is that, if the user wants
to suppress error printing, they can do so externally (e.g. by wrapping the ATC
call with a suitable `with-output`). This avoids the need for inspecting whether
the `:print` input is `nil` or not (in order to tell `make-event-terse` whether
to suppress error output or not) right at the beginning of the ATC
implementation, even before starting input processing proper, which has always
been a bit inelegant. Now that code is more streamlined.

Also improved the user doc of the `:print` input to explain the kinds of ACL2
outputs that `:print` controls.

This slightly modified printing approach should be probably used for other event
macros.