Avoid a constant value in source code that was causing problems for GCL 2.6.14.
The problem seems to have been with the call,
(make-sysfile :system "projects/apply/base.lisp").
The fix is for make-sysfile to call cons explicitly, rather than to
create a sysfile constant when its arguments are constants.  Here is a
new comment in make-sysfile that gives a (not terribly satisfying)
explanation.

; At one time we tried to save conses by returning `(quote (,key . ,str)) when
; (and (keywordp key) (stringp str)).  But with GCL 2.6.14 we got an error from
; confirm-apply-books during certification of "projects/apply/base.lisp",
; perhaps because of the call there of make-sysfile -- that may have created a
; code constant that is destructively modified somehow (e.g., by
; replace-project-dir-alist), though we haven't confirmed that.  At any rate,
; the error went away when we removed that special handling for (and (keywordp
; key) (stringp str)).