[C shallow] Improve some function names.
The conversions functions

 `<type>-integer-list-from-<type>-list`
 `<type>-list-from-<type>-integer-list`

have been renamed to

 `integer-list-from-<type>-list`
 `<type>-list-from-integer-list`

where `<type>` is the fixtype name of each supported C integer type.

That is, `<type>-integer-list` has been abbreviated to `integer-list`, given
that the rest of the name makes it clear what the conversion, and its inputs and
outputs, are. Note that this is consistent with the recent renaming of the
single (not list) integer conversions `<type>-from-integer` and
`integer-from-<type>`: we just add `list` to both sides of the `from` to obtain
the new names that this commit produces.