[ATC] Generalize symbol table information.
In order to support access to global scalar variables, we need to track whether
a variable is global (i.e. an externally declared object).

Adapt the code to properly generate, track, and use this information. In
particular, we use this information to ensure that no side effects on globals
are lost, i.e. that if a function modified a global variable somewhere, it must
return it.