[Std/typed-lists] Add a rule and rename another rule.
The new rule backchains from `true-listp` to `symbol-listp` without limit; it is
disabled by default.

The renamed rule is one that backchains at most once, and is enabled.

It seems that using the unqualified `true-listp-when-symbol-listp-rewrite` for
the first rule is more consistent with other libraries. This is the reason for
renaming the second rule in a way that describes its specific characteristics.