[C deep] Rephrase `exec-ident`.
The definition is now in terms of first finding the object designator of the
variable and then reading the designated object, instead of using `read-var`,
which will be probably removed eventually.

The definition still performs a (partial) array-to-pointer conversion, but this
will be refactored and generalized soon.

Adjust the symbolic execution rule to still work the same as before, so that
proof generation does not need to be modified yet.