[PFCS] Improve formulation of semantics.
Instead of the whole extended assignment for the subtrees, just store the delta
(over the assignment of arguments to parameters) in the proof tree.

Also simplify the proof support rule for relational constraints: now one just
has to exhibit an assignment for the auxiliary/internal variables, not for all
the variables of the relation's body.

This is just a formulation change. There is no conceptual change in the
semantics of PFCSes.