[ATC] Disable some loop stoppers.
All the update ordering rules on the symbolic execution states need to avoid
loop stoppers, because convergence is guaranteed by the ordering, independently
from the size of the involved terms. While the older rules disabled loop
stoppers already, the newer rules failed to do this.

Thanks to Grant Jurgensen for running ATC on an example where proof generation
failed due to the (unwanted) loop stoppers.

Thanks to Matt Kaufmann for reminding me of loop stoppers.