[C deep] Extend/improve dynamic semantics.
Avoid array-to-pointer conversion within the execution of identifiers, since
this conversion is now performed by callers as applicable. This makes the
execution of identifiers not only simpler, but also more general.

Adjust the symbolic execution rule for executing identifiers accordingly.

Previous commits, which extended the symbolic execution theory, also serve to
keep things working with the change in this commit.