[C deep] Improve dynamic semantics.
Adjust `exec-arrsub-of-memberp` to be fully consistent with the combination of
`exec-memberp` and `exec-arrsub`. The difference is irrelevant to our use of the
function in the ATC proofs, but the adjustment will let us simplify the dynamic
semantics, removing this function from it, analogoulsy to what was done for
`exec-arrsub-of-member` in recent commits.