[C deep] Improve dynamic semantics.
Use `exec-expr-pure` to evaluate the left-hand side of an assignment when it is
an identifier, and use `write-object` with the object designator. This moves
things towards the more general formulation of the execution of assignments.

Adapt the symbolic execution rule for assignments to identifiers. We need an
extra hyp (discharged during the symbolic execution), and we need to modify the
proof of the rule slightly.