Avoid duplicate entries in type-alists.  Mentioned cert.pl in :DOC certify-book.
Quoting :DOC note-8-6:

  Duplicate entries in [type-alist]s (proof contexts) are now avoided
  in many cases.  (Implementation note: some calls extending the
  type-alist with an existing term/type-set pair are now avoided in
  source function assume-true-false-rec.)  Thanks to Eric Smith for
  pointing out that there can be type-alists with many consecutive
  identical entries.

Note on testing.  I initially ran regression-everything after making a
change to extend-type-alist-simple so that it causes an error when
attempting to add a term/type-set entry that's already at the front of
the type-alist.  I observed many errors from that before installing
the change to assume-true-false-rec.  Then I ran a fresh
regression-everything, which passed even with those changes to both
functions.  Finally, I reverted to the original definition of
extend-type-alist-simple, keeping only the changes to
assume-true-false-rec, and successfully ran a final, fresh
regression-everything.

Thanks to Bob Boyer for a conversation leading to mention of cert.pl
in :DOC certify-book.