[C deep] Simplify dynamic semantics.
Eliminate the special case `<var>.<member> = <expression>` in the execution of
assignments. This is now handled by the general case.

Adapt the generators of struct-specific member writing theorems. This took a bit
of work, partly because these theorems are generated with hints that involve
quoted theories to avoid "random" failures.