[ATC] Rename some functions generated by `defstruct`.
This renames
```
struct-<tag>-<member>-index-okp ==> struct-<tag>-<member>-integer-index-okp
```

That is, it adds `integer` to make it more clear that this function takes an
ACL2 integer as index.

This frees up the old name for a new version of the function that takes any C
integer as index, which will replace the ones that are specific to the various C
integer types for the indices.