[ATC] Improve theorem generation.
Generate new struct-specific symbolic execution theorems for the recently
introduced operations to write elements of array members with indices of any C
integer type. Use these theorems to speed up the proofs of the similar theorems
for the operations that take specific C integer types for indices.

This cuts significantly (almost to 1/2) the time to process ATC calls that
generate these theorems.

This is for assignments via the `.` operator. A similar improvement will be made
shortly for assignments via the `->` operator.