[FTY] Add a theorem about osets of symbols.
This is useful when the oset is used as a list. It is disabled because normally
one should adhere to the abstraction that osets are not lists. Perhaps there
should be an identity operation that "converts" osets into lists, but for now we
just have a theorem of this form (just for osets/lists of symbols).