[ATC] Deprecate old array operations.
The array operations that are specific to C integer types for the indices are
now deprecated. In order to use them, one has to pass the option `:deprecated
(:arrays)` to ATC. Such options have been added to calls of ATC outside the
community books, so that we can transition away from the deprecated operations
in a piecewise manner, given that there is some material outside the community
books that depends on the deprecated operations -- that material must be
migrated to the new operations.

The `:deprecated` option of ATC is a more general mechanism by which we can
deprecate features and avoid their use in a more controlled way than having to
modify large amounts of code at once. It takes a list of keywords (currently
just `:arrays` is allowed), each of which describes a deprecated feature. We
will soon use this also for operations on structs involving specific C integer
types for indices of array members. The `:deprecated` option is intentionally
not documented in the user doc of ATC, because it should be used just
transitorily to facilitate code migration.