[ATC] Fine-tune the symbolic execution theory more.
Selectively use either the rules for the new struct operations or the rules for
the old struct operations, based on whether `:deprecated` includes `:structs` or
not, both for efficiency and to avoid potential interference. Also do not
generate the rules for the old struct operations at all, unless `:deprecated`
includes `:structs`. (We still need the rules for the new operations when
`:deprecated` includes `:structs` because those rules are proved from the ones
for the new struct operations.)