[C deep] Simplify dynamic semantics.
Eliminate the last special treatment of assignment, for the case `s->m[i] =
...`. Adapt the generated hints for the struct-specific symbolic execution rules
accordingly.

Also add a missing array-to-pointer conversion, on the right-hand side
expression of the assignment. While this does not affect the code currently
generated by ATC, it is important to have an accurate semantics, for more
general uses. This change necessitated a slight adjustment of the proofs
generated for the struct-specific symbolic execution rules for assignments of
the form `s.m[i] = ...`.