[C deep] Move some code away from deep embedding.
The functions to read and write variables, and auxiliary functions, as well as
some theorems about them, are now in a separate file under the `atc`
subdirectory. The definition of the dynamic semantics no longer needs them,
because it uses more general functions that involve object designators. But the
moved functions are still used by ATC, and so they are now under ATC proper (not
the C deep embedding). These may be superseded and removed altogether in the
future.