Further improved interaction between brr and iprint.  Added :DOC for acl2-doc G command.
The previous fix in commit 69912ed4ef was not adequate.  This commit
remedies that and also adds comments, especially at the end of the
Essay on Iprinting, describing how iprinting is made to interact
properly with wormholes.  Here is the replacement release note,
which references new tests.

  Arranged that [iprinting] that takes place during [break-rewrite] is
  better reflected outside break-rewrite.  For examples, see
  [community-books] input file
  books/system/tests/iprint-and-brr-input.lsp, which contains
  comments on what went wrong in Version 8.5, and which generates
  (via the [run-script] utility) the output in file
  iprint-and-brr-log.txt in that directory.

Thanks to J for helpful discussions on this brr/iprinting issue.

Thanks to Warren Hunt for a query leading to documentation for the
(existing) G command in acl2-doc.  (Also, the emacs-level
documentation was fixed; it had failed to mention that G brings up a
new buffer.)