Fixed a bug in with-cbd that caused errors when trying to make include-book forms absolute, as is done for portcullis commands and make-event expansions.  Added analogous utility, with-current-package, for binding the current package.
For the bug fix I added the following to the release note for with-cbd
in :DOC note-8-6.

 Thanks to Sol Swords and Mertcan Temel for reporting a bug in an
 initial implementation.

As part of that fix: in an event context, the first argument of
with-cbd must now be a string, not merely an expression that evaluates
to a string.  This requirement is documented in :DOC with-cbd
(similarly for with-current-package).

Quoting :DOC note-8-6:

  The new utility, [with-current-package], evaluates a given form with
  respect to an indicated [current-package].  Thanks to Sol Swords
  for requesting this utility.

NOTE 1:
Currently with-current-package can only be used in :program mode.  I
may remedy that soon, and at that time (if there's no objection) I
expect that I'll add with-current-package to *acl2-exports*.

NOTE 2:
Sol Swords suggested (5/18/2023) that we consider a general macro that
binds a given state global; then with-cbd and with-current-package
would be macros that lay down calls of that general macro.  I think
this will be pretty trivial, except that it might be necessary to
disallow certain state globals, or allow special code (see the
discussion in with-current-package about how it's simpler than
with-cbd).  Rather than think that through now, I've decided just to
keep with-cbd and with-current-package as independent primitives.  If
someone has a use for the general macro described above, we can talk.