[ATC] Extend modular proof generation.
Actually enable modular theorems for the initial scope involving pointers to
integers, by moving and modifying the code to set the flag to stop modular proof
generation, and by using a separate flag for these theorems.

Fix the formula of the generated theorem by taking into acount the `-ptr`
variables.