[ATC] Extend modular proof generation.
Handle the indirection operator (unary `*`), for reading.

This necessitated the extension of a theorem generation utility to take two
possibly different terms for the two parts of the generate theorem; these are
often equal, but in the case of an expression that is a pointer variable the two
terms have to differ, one being `var-ptr` and the other being `var`.