[C deep] Adjust `signed char` min.
If the signed representation is two's complement but the pattern with sign bit 1
and all value bits 0 is a trap representation, then the minimum value is
`-(2^(CHAR_BIT-1)-1)` (e.g. 127 for 8-bit bytes), not `-2^(CHAR_BIT-1)`
(e.g. 128 for 8-bit bytes). This is only used in the more general definition of
implementation environments, which the rest of the C formalization does not use
yet.