Fixed bug in acl2-doc searching, again closing GitHub Issue 1487.
This bug was introduced with yesterday's commit to render text
documentation to respect fonts by default.  I closed GitHub Issue 1487
at that time, but then reopened it when discovering that generated
file acl2-doc-search has SGR sequences in it.  That's what I fixed;
that file no longer has SGR sequences in it.

I plan to update the manual on the web shortly.  In the meantime, to
get the effect of this fix you would need to delete file
books/system/doc/acl2-doc-search and optionally rebuild it by
certifying books/top.