[PFCS] Improve names in abstract syntax.
Use strings instead of symbols, for relation and variable names.

This is more streamlined, as there are no extra package names (which may depend
on the packages' imports), and it is also more consistent with the concrete
syntax, which uses identifiers for these names.

Extend the lifter to map strings to symbols according to typical conventions:
turn lowercase into uppercase, and underscore into dash.

Adapt other parts of the PFCS library as needed.