[ATC] Improve `defstruct`.
Avoid the `deftheory`, and its use in some of the proofs, by using quoted
theories for all the generated proofs. This was previously done for most proofs,
but not all; now it's all of them. Also generate `:guard-simplify :limited` to
avoid an unwanted dependency on guard simplification (which is effectively part
of guard verification when a function is submitted, although technically guard
verification is a separate thing) on the current theory.