[ATC] Remove redundant check.
Structs with flexible array members can only be passed by pointer in ATC, so if
we get to the point of generating a modular proof for an expression that reads
an element of an array member of a struct passed by value, there is no need to
check that the array is not flexible. This check is already performed elsewhere.