[ATC] Add missing check to `defstruct`.
If there is a flexible array member, there must be at least another member. This
check was missing from `defstruct`, but it has now been added. The issue was
discovered in the course of testing the generation of modular proofs for
structs: an example struct with just one flexible array member (and no other
member) caused a failure in the generated well-formedness theorem, because our
formal static semantics of C correctly enforces that requirements on structs
with flexible array members.