Writing of event-data files now takes place with "ACL2" as the current-package.  Those files and useless-runes files are now read with "ACL2" as the current-package.  Deleted a false claim in :DOC set-raw-mode about being put into the Lisp debugger.
Quoting :DOC note-8-6:

  Fixed the [useless-runes] feature to work properly when reading a
  useless-runes file while certifying a book in a package other than
  the "ACL2" package.  The fix is to ensure that the useless-runes
  file is read while in the "ACL2" package, which is the same package
  used when the file was written.

Thanks to Eric Smith for requesting that files *@event-data.lsp be
written while in the ACL2 package, just as is done for
*@useless-runes.lsp files.  While making that change, I discovered
that *@useless-runes.lsp files were being read in the package that is
the current-package when certify-book is called, rather than the
"ACL2" package.  I fixed that as indicated in the release note above.
And of course, *@event-data.lsp are now also read with the
current-package being the "ACL2" package.

Thanks to Eric McCarthy for pointing out the bug in :DOC set-raw-mode.
(One can get into the Lisp debugger in raw-mode just as one normally
does, by using set-debugger-enable.)