[ATC] Refactor some code.
Have the code to contextualize compustation states take two contexts (the second
extending the first) instead of a list of premises corresponding to the
difference between the two contexts. This avoids some code replication, and it
also performs additional internal validation checks by ensuring that the two
contexts have identical preambles, in addition to one having a prefix of the
premises of the other.