[ATC] Refine a test.
This is actually enabling modular proof generation to continue after a `let`
that binds a single variable to a side-effecting `if`. This was claimed in the
previous commit, but the test was too strong. The test has been relaxed, but it
has also been strengthened along another dimension, namely to stop modular proof
generation in this situation when we are in a loop (this will be supported
later), which was otherwise causing a failure in an example outside the
community books.