[ATC] Improve user interface.
When `:print` is at least `:result`, print the `defconst` event just after it is
accepted, instead of printing it at the end. This is part of an improvements in
which the results are printed as they are submitted, which gives a better sense
of progress even with the default `:print`.