[ATC] Extend modular proof generation.
Generate modular proofs for not only array assignments and their subsequent
code, but also for their combination. This lets modular proofs proceed, for
certain functions, all the way to the point in which the frame is popped, i.e.
almost at the end of the function execution. Add a check to stop modular proofs
at that point if there are net side effects in that function, because those need
a further (upcoming) extension to modular proof generation, in particular the
theorem about popping the frame.