[ABNF] Put PDF example in the ABNF package.
This is as done in the other examples in the same directory, and avoids some
package prefixes. Since it is just a stand-alone example for now, it does not
matter much which package it is in. If this becomes part of a larger PDF
development, probably this file would be moved there, and perhaps put into a
`PDF` package.