Write bookdata file based on environment variable ACL2_WRITE_BOOKDATA.  Files BOOK@event-data.lsp are now written to the directory of BOOK, not to the .sys/ subdirectory.
Quoting :DOC note-8-6:

  When environment variable ACL2_WRITE_BOOKDATA has a non-empty value,
  then files of the form *__bookdata.out will be written; see
  [bookdata] for explanation and an exception.  Thanks to Eric Smith
  for requesting this enhancement.

Thanks to Eric Smith for also requesting the event-data change.  The
.sys/ directories created for event-data files were showing up in "svn
status" commands.  This change also follows the convention for
*@expansion.lsp files: in both cases, the file is not in the
repository and is in the same directory as the book.  There is no new
release note about this change to writing of event-data, since that is
not in ACL2 Version 8.5.