[omaps] Disable a rule.
This breaks the abstraction between omaps and alists, so it should be disabled,
at least by default. Apparently some proofs in the file depend on it, so it is
enabled in the proofs that need it. Perhaps in the future we may rephrase even
those proofs in a way that doesn't need that rule.