[ATC] Improve some code for modular proofs.
In order to properly handle side effects of multiple objects in modular proofs,
we need to revise the form of the result type assertions in the generated
modular theorems. In the case of two or more results, we need to push `mv-nth`
into certain kinds of terms, but not others. This commit makes this revision.