[APT] Improve `restrict` transformation.
For the recursive case, generate `:expand` hints in the correctness theorem,
even though the generated quoted theory includes the definitions. We observed a
case in which the desired expansion did not occur, presumably due to some ACL2
heuristic. So we add `:expand` hints.

Thanks to Eric Smith for running the `restrict` transformation on an example
that failed as described above.