[APT] Improve `restrict` guard proofs.
To avoid a failure in an example that seems caused by ACL2's special treatment
of `not`, we include `not` in the generated quoted theory for verifying the
guards of the new function. All our currnt uses of `restrict` still works, but,
as mentioned in the added documentation, it is not clear how robust this
approach is.