Fixed markup issues in new :DOC sbcl-installation and added :DOC sbcl-installation-brief.  Synched doc.lisp etc.
Recently Alessandro replaced :DOC sbcl-installation with a new version
that mentions the need to install an SBCL binary first and also adds
helpful details.  However, I've used the old instructions successfully
for some years, so I wanted to preserve those.  We agreed via email on
having two topics: his new one, named sbcl-installation, and the old
one, which I've named sbcl-installation-brief (following the precedent
of existing topics ccl-installation-mac-brief and
ccl-installation-linux-brief).  I've tried to make clear, in the :DOC,
the role of each topic as discussed above.

In the process I noticed markup in the new topic of the form @{(...)},
which was presumably intended to be @({...}).  In fixing these I found
that some punctuation needed adjustment.  (I have no objection to
anyone further modifying the punctuation, of course.)