Support ACL2 builds on older Lisps.
Thanks to Alessandro Coglio for noting that some older versions of CCL
failed to support building saved_acl2, and thanks to Eric Smith for
weighing in positively on my proposed solution.  The build failure was
because ACL2 now insists on the presence of Lisp feature
:ieee-floating-point (in support of the new support for floating-point
computations).  Now, the error message points to a new :DOC topic, fp,
that suggests a workaround using an environment variable, with the
caveat that the Lisp properly supports IEEE floating-point operations.
I tested with "make regression-everything-fresh" on a platform with an
older CCL, by adding the following line to my ccl script.

export ACL2_FP_OK=t