[ABNF] Extend `deftreops`.
For each of two or more alternatives that define a rule name, also generate a
theorem asserting the equivalence between (1) the alternative checking function
returning the positive integer that indicates the alternative and (2) the
discriminating term for that alternative. These theorems are generated as part
of the generated `define` for the alternative checking function.