Remove unneeded acl2-mv-as-values from profiling raw code
Some of the raw Lisp code for misc/profiling was gated on the
acl2-mv-as-values feature, which was removed from ACL2 in August 2021 as
it is true on all of the Lisps that ACL2 supports at this time. This
meant that the raw Lisp code would never evaluate, preventing
misc/profiling from working. This commit modifies the code to stop
gating the relevant code on the acl2-mv-as-values flag.