[X86ISA] Fix bug in `def-inst`.
Attempting to add a VEX instructions resulted in an error because the semantic
function (defined via `def-inst`) was expecting two arguments `vex-prefixes` and
`evex-prefixes`, while the caller was only passing `vex-prefixes`.

This commit fixes the issue by adding a `:vex` option to `def-inst`, to
conplement the existing `:evex` option, so that either `vex-prefixes` or
`evex-prefixes` are generated as formals by `def-inst`.

Also auto-remove some trailing whitespace.