[X86ISA] Fix bug in VEX dispatch generator.
At the point of the code that this commit modifies, `caseStmt` is a list of
conditions. The code was meant to turn that into a single condition, but was
doing so incorrectly in some cases. In particular, given an empty list of
conditions, it was generating `nil` instead of `t`.

The problem manifested itself when attempting to add a VEX instruction: the
generated dispatch code has `cond` clauses whose conditions were `(and
... nil)`, which were therefore always false.

This commit fixes the problem.