[xdoc] Get rid of LazyLoad.js, use fetch instead.
Get rid of the LazyLoad.js library, use JavaScript's built-in fetch
functionality instead. Also replace the use of JQuery's AJAX
functionality with fetch.

Also added a prefetch directive for xindex.js to index.html so that in
theory the browser knows to start loading it before the JavaScript fetch
initiates the request for it.

Also clean up xdataWhenReady, rename it to xdataLoadKeys, and have it
return a promise instead of taking in a callback. This function now uses
fetch instead of JQuery's fetch functionality.