[xdoc] Factor XSLT code into a class, fix local file loading
The XSLT transformer now has its own class in xslt.js, and the place
where the XSL template is loaded was changed.

Add loadJS to support using the fancy viewer when loaded from a file,
without a HTTP server. Fetch and most modern JS request APIs will not
work without a HTTP server due to CORS issues.

Use loadJS to load the xindex and xdata files, in addition to the
render.js file.

Remove jquery.base64 as it is unneeded on all major browsers post-2010.
(instead we use the native JavaScript atob function)