[X86ISA] Add functions to read/write memory.
These are for 256 and 512 bits, needed by some upcoming instructions that move
data between memory and YMM/ZMM registers.

Two functions for 512 bits work but are very slow to certify, so they are
commented out for now, as indicated in comments just before them. We do not need
them for VEX-encoded instructions, but we will need them for EVEX-encoded
instructions.

Also adjust x86 reasoning rules in the Kestrel Books.