[X86ISA] Generalize function to retrieve operands.
This is the main function that retrieves operands from registers or memory.
This commit generalizes it to support access to XMM and YMM registers. More
precisely, the access to XMM registers is for VEX instructions; the function
already had support for accessing XMM registers for non-VEX instructions. (The
two access modes are slightly different.)