Fixed buggy type declarations in compress2.
Thanks to Alessandro Coglio for reporting a build failure on ARM and
letting me know which commit caused the problem.  That was using SBCL,
but I was able to reproduce an error building with safety 3 and CCL on
a Mac.  That error is gone now, so I'm optimistic that the ARM build
will succeed now.