Made redefinitions of char-upcase and string-upcase in CMUCL, and string-downcase in SBCL, to work around Lisp bugs.
The relevant code is below the following comments in ACL2 source file
acl2.lispl.

; Fix CMUCL bug through April 2024.
; Fix SBCL bug, fixed in SBCL on 6/2/2024.

I made the CMUCL mod previously but I've improved it (no longer
causing an error, and no longer redefining functions when that isn't
necessary, probably starting with the next CMUCL snapshot release).
The SBCL mod is new, but that SBCL bug was fixed on 6/2/2024 after
very soon after I reported it.

Quoting a comment in (defxdoc note-8-6 ...):

; Redefined char-upcase and string-upcase in CMUCL, and string-downcase in
; SBCL, when necessary to avoid Lisp bugs reported for CMUCL and SBCL in,
; respectively, https://gitlab.common-lisp.net/cmucl/cmucl/-/issues/316 and
; https://bugs.launchpad.net/sbcl/+bug/2067841.