[C$] Fix typos in lexer.
The function to lex escape sequences repeated the same test instead of testing
for the different letters of the escapes, resulting in escapes with those
letters not being recognized.

Thanks to Eric Smith for running the parser on an example that led to the
discovery and fix of this issue.