[C$] Extend abstract syntax.
Capture what should be all forms of ambiguous cast expressions in the fixtype
for expressions.

This adds a case to the fixtype, and extends the existing four cases.

Also adds detailed discussion of those ambiguities.

Adapt dependents to certify. Next, the parser needs to be extended to take
advantage of the extended ASTs.