[C$] Extend parser to handle GCC attributes.
This only handles certain forms and occurrences of GCC attributes, as detailed
in the grammar and abstract syntax.

This commit adds functions to parse attributes, and extends certain existing
functions to conditionally attempt to recognize and handle attributes, when the
GCC flag in the parser state is set. This also interacts with the code to
disambiguate external declarations.